requirements.txt
playwright==1.52.0
greenlet>=3.2.2
cryptography==41.0.5
beautifulsoup4==4.12.2
requests==2.31.0
ttkthemes==3.2.2

pytest-asyncio
